requirements.txt
anyio==3.6.1
azure-core==1.26.0
azure-storage-blob==12.13.1
boto3==1.24.88
botocore==1.27.88
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.1
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
